+++ my-app/src/components/cardList/CardRow/CardRow.tsx
@@ -1,6 +1,6 @@
 import type { CardListItem } from "../../../apis/cards";
-import { useDeleteCard } from "../../../hooks/queries/useDeleteCard";
 import { ISSUERS, isIssuerCode } from "../../../constants/issuers";
+import DeleteButton from "../DeleteButton/DeleteButton";
 
 type Props = {
   card: CardListItem;
@@ -8,14 +8,7 @@ type Props = {
 };
 
 const CardRow = ({ card, onDelete }: Props) => {
-  const { remove } = useDeleteCard();
   const issuer = isIssuerCode(card.issuerCode) ? ISSUERS[card.issuerCode] : null;
-
-  const handleDelete = async () => {
-    if (!window.confirm("정말 삭제하시겠습니까?")) return;
-    await remove(card.id);
-    onDelete();
-  };
 
   return (
     <li>
@@ -25,7 +18,7 @@ const CardRow = ({ card, onDelete }: Props) => {
         <p>{card.number}</p>
         <p>유효기간 {card.expirationDate}</p>
       </div>
-      <button onClick={handleDelete} aria-label="카드 삭제">×</button>
+      <DeleteButton cardId={card.id} onDelete={onDelete} />
     </li>
   );
 };
